fix: change paths from './createDataContext' to 'lib/utils/createDataContext' because createDataContext file moves to lib/utils
